dashboard/app: show manager unique coverage
1. Make heatmap testable, move out the spanner client instantiation.
2. Generate spannerdb.ReadOnlyTransaction mocks.
3. Generate spannerdb.RowIterator mocks.
4. Generate spannerdb.Row mocks.
5. Prepare spannerdb fixture.
6. Fixed html control name + value.
7. Added multiple tests.
8. Show line coverage from selected manager.
9. Propagate coverage url params to file coverage url.

+++ pkg/cover/file.go
@@ -101,7 +101,7 @@ func GetMergeResult(c context.Context, ns, repo, forCommit, sourceCommit, filePa
 func rendResult(content string, coverage *covermerger.MergeResult, renderConfig *CoverageRenderConfig) string {
 	if coverage == nil {
 		coverage = &covermerger.MergeResult{
-			HitCounts:   map[int]int{},
+			HitCounts:   map[int]int64{},
 			LineDetails: map[int][]*covermerger.FileRecord{},
 		}
 	}
